change plugin helper name to docker, adding docker binds helper, adding new items to API load
